added mapdl geometry and chained commands
