+++ pyansys/mapdl_old.py
@@ -4,18 +4,14 @@ by MapdlCorba and MapdlConsole for interfacing with MAPDL without the
 use of gRPC.
 """
 import glob
-import tempfile
 import os
 import re
-from shutil import rmtree, copyfile
 
 import numpy as np
-import pyvista as pv
-from vtk import vtkAppendPolyData
 
 from pyansys.mapdl import _MapdlCore
 import pyansys
-from pyansys.misc import is_float, supress_logging
+from pyansys.misc import is_float, supress_logging, chunks
 
 
 def check_lock_file(path, jobname, override):
@@ -336,97 +332,6 @@ class _MapdlOld(_MapdlCore):
         """Reset cached items"""
         self._archive_cache = None
 
-    @supress_logging
-    def _load_lines(self):
-        """Load lines from MAPDL using IGES"""
-        # write database to an archive file
-        filename = os.path.join(self.path, 'tmp.iges')
-
-        # ignore volumes
-        self.cm('__tmp_volu__', 'VOLU')
-        self.cm('__tmp_area__', 'AREA')
-        self.cm('__tmp_line__', 'LINE')
-        self.allsel()
-        self.vsel('NONE')
-
-        prior_processor = self.processor
-        self.prep7()
-        self.igesout(filename, att=1)
-        self.run('/%s' % prior_processor)
-
-        self.cmsel('S', '__tmp_volu__', 'VOLU')
-        self.cmsel('S', '__tmp_area__', 'AREA')
-        self.cmsel('S', '__tmp_line__', 'LINE')
-
-        iges = pyiges.read(filename)
-
-        selected_lnum = self.lnum
-        lines = []
-        entity_nums = []
-        for bspline in iges.bsplines():
-            # allow only 10001 as others appear to be construction entities
-            if bspline.d['status_number'] == 10001:
-                entity_num = int(bspline.d['entity_subs_num'])
-                if entity_num not in entity_nums and entity_num in selected_lnum:
-                    entity_nums.append(entity_num)
-                    line = bspline.to_vtk()
-                    line.cell_arrays['entity_num'] = entity_num
-                    lines.append(line)
-
-        entites = iges.lines() + iges.circular_arcs()
-        for line in entites:
-            if line.d['status_number'] == 1:
-                entity_num = int(line.d['entity_subs_num'])
-                if entity_num not in entity_nums and entity_num in selected_lnum:
-                    entity_nums.append(entity_num)
-                    line = line.to_vtk(resolution=100)
-                    line.cell_arrays['entity_num'] = entity_num
-                    lines.append(line)
-
-        if lines:
-            afilter = vtkAppendPolyData()
-            [afilter.AddInputData(line) for line in lines]
-            afilter.Update()
-            self._lines = pv.wrap(afilter.GetOutput())
-        else:
-            lines = pv.PolyData()
-
-    @supress_logging
-    def _load_keypoints(self):
-        """Load keypoints from MAPDL using IGES"""
-
-        # write database to an archive file
-        filename = os.path.join(self.path, 'tmp.iges')
-
-        # write only keypoints
-        self.cm('__tmp_volu__', 'VOLU')
-        self.cm('__tmp_area__', 'AREA')
-        self.cm('__tmp_line__', 'LINE')
-        self.vsel('NONE')
-        self.asel('NONE')
-        self.lsel('NONE')
-
-        prior_processor = self.processor
-        self.prep7()
-        self.igesout(filename, att=1)
-        self.run('/%s' % prior_processor)
-
-        self.cmsel('S', '__tmp_volu__', 'VOLU')
-        self.cmsel('S', '__tmp_area__', 'AREA')
-        self.cmsel('S', '__tmp_line__', 'LINE')
-
-        iges = pyiges.read(filename)
-
-        keypoints = []
-        kp_num = []
-        for kp in iges.points():
-            keypoints.append([kp.x, kp.y, kp.z])
-            kp_num.append(int(kp.d['entity_subs_num']))
-
-        # self._kp_num = np.array(self._kp_num)
-        self._keypoints = pv.PolyData(keypoints)
-        self._keypoints['entity_num'] = kp_num
-
     @property
     @supress_logging
     def _mesh(self):
@@ -438,61 +343,6 @@ class _MapdlOld(_MapdlCore):
             self._archive_cache = pyansys.Archive(arch_filename, parse_vtk=True,
                                                   name='Mesh')
         return self._archive_cache
-
-    # def _list(self, command):
-    #     """ Replaces *LIST command """
-    #     items = command.split(',')
-    #     filename = os.path.join(self.path, '.'.join(items[1:]))
-    #     if os.path.isfile(filename):
-    #         self._response = open(filename).read()
-    #         self._log.info(self._response)
-    #     else:
-    #         raise Exception('Cannot run:\n%s\n' % command + 'File does not exist')
-
-    def eplot(self, vtk=False, **kwargs):
-        """APDL Command: EPLOT
-
-        Produces an element display.
-
-        Parameters
-        ----------
-        vtk : bool, optional
-            Plot the currently selected elements using ``pyvista``.
-
-        **kwargs
-            See ``help(pyvista.plot)`` for more keyword arguments
-            related to visualizing using ``vtk``.
-
-        Notes
-        -----
-        Produces an element display of the selected elements. In full
-        graphics, only those elements faces with all of their
-        corresponding nodes selected are plotted. In PowerGraphics,
-        all element faces of the selected element set are plotted
-        irrespective of the nodes selected.  However, for both full
-        graphics and PowerGraphics, adjacent or otherwise duplicated
-        faces of 3-D solid elements will not be displayed in an
-        attempt to eliminate plotting of interior facets. See the DSYS
-        command for display coordinate system issues.
-
-        This command will display curvature in midside node elements
-        when PowerGraphics is activated [/GRAPHICS,POWER] and
-        /EFACET,2 or /EFACET,4 are enabled.  (To display curvature,
-        two facets per edge is recommended [/EFACET,2]).  When you
-        specify /EFACET,1, PowerGraphics does not display midside
-        nodes. /EFACET has no effect on EPLOT for non-midside node
-        elements.
-
-        This command is valid in any processor.
-        """
-        if vtk:
-            # default kwargs
-            kwargs.setdefault('color', 'w')
-            kwargs.setdefault('show_axes', True)
-            kwargs.setdefault('show_edges', True)
-            return self._vtk_grid.plot(**kwargs)
-        else:
-            return super().eplot(**kwargs)
 
     def load_parameters(self):
         """Loads and returns all current parameters
@@ -571,248 +421,6 @@ class _MapdlOld(_MapdlCore):
     def _get(self, *args, **kwargs):
         """Simply use the default get method"""
         return self.get(*args, **kwargs)
-
-    def get(self, par="__floatparameter__", entity="", entnum="",
-            item1="", it1num="", item2="", it2num="", **kwargs):
-        """APDL Command: *GET
-
-        Retrieves a value and stores it as a scalar parameter or part
-        of an array parameter.
-
-        Parameters
-        ----------
-        par
-            The name of the resulting parameter. See *SET for name
-            restrictions.
-
-        entity
-            Entity keyword. Valid keywords are NODE, ELEM, KP, LINE, AREA,
-            VOLU, PDS, etc., as shown for Entity = in the tables below.
-
-        entnum
-            The number or label for the entity (as shown for ENTNUM = in the
-            tables below). In some cases, a zero (or blank) ENTNUM represents
-            all entities of the set.
-
-        item1
-            The name of a particular item for the given entity. Valid items are
-            as shown in the Item1 columns of the tables below.
-
-        it1num
-            The number (or label) for the specified Item1 (if any). Valid
-            IT1NUM values are as shown in the IT1NUM columns of the tables
-            below. Some Item1 labels do not require an IT1NUM value.
-
-        item2, it2num
-            A second set of item labels and numbers to further qualify the item
-            for which data are to be retrieved. Most items do not require this
-            level of information.
-
-        Returns
-        -------
-        par : float
-            Floating point value of the parameter.
-
-        Examples
-        --------
-        Retreive the number of nodes
-
-        >>> value = ansys.get('val', 'node', '', 'count')
-        >>> value
-        3003
-
-        Retreive the number of nodes using keywords.  Note that the
-        parameter name is optional.
-
-        >>> value = ansys.get(entity='node', item1='count')
-        >>> value
-        3003
-
-        Notes
-        -----
-        *GET retrieves a value for a specified item and stores the value as a
-        scalar parameter, or as a value in a user-named array parameter. An
-        item is identified by various keyword, label, and number combinations.
-        Usage is similar to the *SET command except that the parameter values
-        are retrieved from previously input or calculated results. For example,
-        *GET,A,ELEM,5,CENT,X returns the centroid x-location of element 5 and
-        stores the result as parameter A. *GET command operations, along with
-        the associated Get functions return values in the active coordinate
-        system unless stated otherwise. A Get function is an alternative in-
-        line function that can be used to retrieve a value instead of the *GET
-        command (see Using In-line Get Functions for more information).
-
-        Both *GET and *VGET retrieve information from the active data stored in
-        memory. The database is often the source, and sometimes the information
-        is retrieved from common memory blocks that the program uses to
-        manipulate information. Although POST1 and POST26 operations use a
-        *.rst file, *GET data is accessed from the database or from the common
-        blocks. Get operations do not access the *.rst file directly. For
-        repeated gets of sequential items, such as from a series of elements,
-        see the *VGET command.
-
-        Most items are stored in the database after they are calculated and are
-        available anytime thereafter. Items are grouped according to where they
-        are usually first defined or calculated. Preprocessing data will often
-        not reflect the calculated values generated from section data. Do not
-        use *GET to obtain data from elements that use calculated section data,
-        such as beams or shells. Most of the general items listed below are
-        available from all modules. Each of the sections for accessing *GET
-        parameters are shown in the following order:
-
-        *GET General Entity Items
-
-        *GET Preprocessing Entity Items
-
-        *GET Solution Entity Items
-
-        *GET Postprocessing Entity Items
-
-        *GET Probabilistic Design Entity Items
-
-        The *GET command is valid in any processor.
-        """
-        command = "*GET,%s,%s,%s,%s,%s,%s,%s" % (str(par),
-                                                 str(entity),
-                                                 str(entnum),
-                                                 str(item1),
-                                                 str(it1num),
-                                                 str(item2),
-                                                 str(it2num))
-        response = self.run(command, **kwargs)
-        try:
-            return float(response.split('=')[-1])
-        except:
-            return None
-
-    def read_float_parameter(self, parameter_name):
-        """Read out the value of a ANSYS parameter to use in python.
-
-        Parameters
-        ----------
-        parameter_name : str
-            Name of the parameter inside ANSYS.
-
-        Returns
-        -------
-        float
-            Value of ANSYS parameter.
-
-        Examples
-        --------
-        >>> ansys.get('myparm', 'node', '', 'count')
-        >>> value = ansys.read_float_parameter('myparm')
-        >>> value
-        3003
-
-        Retreive the value in an array.  This example creates a block
-        of elements and stores the element numbers in an array.  Then
-        it retreives the first value of that array
-
-        >>> ansys.prep7()
-        >>> ansys.block(0,1,0,1,0,1)
-        >>> ansys.et(1, 186)
-        >>> ansys.vmesh('all')
-        >>> ansys.run('*VGET, ELEM_ARR, ELEM, , elist')
-        >>> ansys.read_float_parameter('ELEM_ARR(1)')
-        1.0
-
-        You could also retrieve ELEM_ARR with
-
-        >>> values, arrays = ansys.load_parameters()
-        >>> arrays
-        {'ELEM_ARR': array([1., 2., 3., 4., 5., 6., 7., 8.])}
-
-        """
-        try:
-            response = self.run(parameter_name + " = " + parameter_name)
-        except TypeError:
-            raise TypeError('Input variable `parameter_name` should be string')
-        return float(response.split('=')[-1])
-
-    def read_float_from_inline_function(self, function_str):
-        """Use a APDL inline function to get a float value from ANSYS.
-        Take note, that internally an APDL parameter __floatparameter__ is
-        created/overwritten.
-
-        Parameters
-        ----------
-        function_str : str
-            String containing an inline function as used in APDL..
-
-        Returns
-        -------
-        float
-            Value returned by inline function..
-
-        Examples
-        --------
-        >>> inline_function = "node({},{},{})".format(x, y, z)
-        >>> node = apdl.read_float_from_inline_function(inline_function)
-        """
-        self.run("__floatparameter__="+function_str)
-        return self.read_float_parameter("__floatparameter__")
-
-    def open_gui(self, include_result=True):
-        """Saves existing database and opens up APDL GUI
-
-        Parameters
-        ----------
-        include_result : bool, optional
-            Allow the result file to be post processed in the GUI.
-        """
-
-        # specify a path for the temporary database
-        temp_dir = tempfile.gettempdir()
-        save_path = os.path.join(temp_dir, 'ansys_tmp')
-        if os.path.isdir(save_path):
-            rmtree(save_path)
-        os.mkdir(save_path)
-
-        name = 'tmp'
-        tmp_database = os.path.join(save_path, '%s.db' % name)
-        if os.path.isfile(tmp_database):
-            os.remove(tmp_database)
-
-        # get the state, close, and finish
-        prior_processor = self.processor
-        self.finish()
-        self.save(tmp_database)
-        self.exit(close_log=False)
-
-        # copy result file to temp directory
-        if include_result:
-            resultfile = os.path.join(self.path, '%s.rst' % self.jobname)
-            if os.path.isfile(resultfile):
-                tmp_resultfile = os.path.join(save_path, '%s.rst' % name)
-                copyfile(resultfile, tmp_resultfile)
-
-        # write temporary input file
-        start_file = os.path.join(save_path, 'start%s.ans' % self.version)
-        with open(start_file, 'w') as f:
-            f.write('RESUME\n')
-
-        # some versions of ANSYS just look for "start.ans" when starting
-        other_start_file = os.path.join(save_path, 'start.ans')
-        with open(other_start_file, 'w') as f:
-            f.write('RESUME\n')
-
-        # issue system command to run ansys in GUI mode
-        cwd = os.getcwd()
-        os.chdir(save_path)
-        os.system('cd "%s" && "%s" -g -j %s' % (save_path, self._exec_file, name))
-        os.chdir(cwd)
-
-        # must remove the start file when finished
-        os.remove(start_file)
-        os.remove(other_start_file)
-
-        # reload database when finished
-        self._launch()
-        self.resume(tmp_database)
-        if prior_processor is not None:
-            if 'BEGIN' not in prior_processor:
-                self.run('/%s' % prior_processor)
 
     def load_array(self, arr, name):
         """Load a numpy array or python list directly to MAPDL
